#20 [rename] school의 domain package rename
