+++ src/main/java/org/crops/fitserver/domain/school/domain/School.java
@@ -1,4 +1,4 @@
-package org.crops.fitserver.domain.school.entity;
+package org.crops.fitserver.domain.school.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
