+++ src/main/java/org/crops/fitserver/domain/school/repository/SchoolRepository.java
@@ -1,7 +1,7 @@
 package org.crops.fitserver.domain.school.repository;
 
 import java.util.List;
-import org.crops.fitserver.domain.school.entity.School;
+import org.crops.fitserver.domain.school.domain.School;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SchoolRepository extends JpaRepository<School, Long> {
