+++ src/test/java/org/crops/fitserver/domain/school/repository/SchoolRepositoryTest.java
@@ -3,7 +3,7 @@ package org.crops.fitserver.domain.school.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import lombok.RequiredArgsConstructor;
-import org.crops.fitserver.domain.school.entity.School;
+import org.crops.fitserver.domain.school.domain.School;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
